chore: refactor sponsoring component
